Separating out v1 BigtableAdmin into a separate class. This is in preparation for hbase v2. This is mostly a straight refactoring of the exiting implementation of AbstractBigtableAdmin from BigtableConnection.  While I was at this, I refactored the AbstractBigtableAdmin constructor.
